split out mixpanel code
